Remove mappings of the form id -> empty because they're artifacts and shouldn't make it to the data
Thomas Maurel from biomart-users:
This is an expected behaviour of the BioMart software. The GO xrefs are attached to Transcripts in the Ensembl database so if you only ask for Gene in your mart query, you will get one GO result per transcript so you will see ENSMUSG00000000085 multiple times with empty line if the Transcript don’t have a GO term attached to it.